+++ Heatmap/Range/AdaptiveRangeFactory.cs
@@ -4,6 +4,6 @@ namespace Heatmap.Range
 {
     public class AdaptiveRangeFactory : IRangeFactory
     {
-        public IRange Create(IEnumerable<float> values) => new AdaptiveRange(values);
+        public IRange Create(IEnumerable<double> values) => new AdaptiveRange(values);
     }
 }
